// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
  MODERATOR: 'MODERATOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TwoFactorMethod = {
  SMS: 'SMS',
  AUTHENTICATOR: 'AUTHENTICATOR'
} as const

export type TwoFactorMethod = (typeof TwoFactorMethod)[keyof typeof TwoFactorMethod]


export const NewsType = {
  VIDEO: 'VIDEO',
  SHORT: 'SHORT',
  PHOTO: 'PHOTO',
  READ: 'READ'
} as const

export type NewsType = (typeof NewsType)[keyof typeof NewsType]


export const ContentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ContentStatus = (typeof ContentStatus)[keyof typeof ContentStatus]


export const AdStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED'
} as const

export type AdStatus = (typeof AdStatus)[keyof typeof AdStatus]